requirements.txt
keyboard==0.13.5
mediapipe==0.10.20
PyAudio==0.2.14
pytest==8.3.4
Requests==2.32.3
roslibpy==1.8.0
setuptools==75.1.0
sparkai==0.3.0
websocket_client==1.8.0
